requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
djangorestframework-simplejwt
django-cors-headers
uuid
python-decouple